\name{basemean, coefmean}
\alias{basemean}
\alias{coefmean}
\title{
The posterior mean of base or coefficient matrix
}
\description{
\code{basemean} returns the posterior mean of base matrix.
\code{coefmean} returns the posterior mean of coefficient matrix.
}
\usage{
basemean(obj)

coefmean(obj)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{obj}{
  output object of \code{vb_nmf_pois} or \code{svb_nmf_pois}
}
}
\seealso{
\code{\link{vb_nmf_pois}}
}
\examples{
#see example in `?vb_nmf_pois`
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory (show via RShowDoc("KEYWORDS")):
% \keyword{ ~kwd1 }
% \keyword{ ~kwd2 }
% Use only one keyword per line.
% For non-standard keywords, use \concept instead of \keyword:
% \concept{ ~cpt1 }
% \concept{ ~cpt2 }
% Use only one concept per line.
