\name{VBsNMF-package}
\alias{VBsNMF-package}
\alias{VBsNMF}
\docType{package}
\title{
\packageTitle{VBsNMF}
}
\description{
\packageDescription{VBsNMF}
}
\details{
%%  ~~ An overview of how to use the package, ~~
%%  ~~ including the most important functions ~~
}
\author{
\packageAuthor{VBsNMF}

Maintainer: \packageMaintainer{VBsNMF}
}
\keyword{package}

%% Uncomment below to imitate parts of library(help = ssnmf)
%\section{The \file{DESCRIPTION} File}{\packageDESCRIPTION{ssnmf}}
%\section{Documentation Index}{\packageIndices{ssnmf}}

\examples{
%%  ~~ Optional simple examples of the most important functions ~~
}
